More natural car rotation
Changed car physics so instead of getting one blocks colliding box, it
gets 3 blocks in a line (rotated ofc). This way, if there's a one wide
hole, the car won't be able to dip into it. This makes for a much more
natural feel of the car, as movement points of the car are more spread
out, rather than in one position.